- fixed network graphic - debuged evaluation tables - changed cache settings in template engine - some speed hacks - changed int angles for peer positions in network graphic to double angles
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@8124 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/net/yacy/visualization/PrintTool.java
@@ -110,7 +110,7 @@ public class PrintTool {
 
 
     private static final int arcDist = 8;
-    public static void arcPrint(final RasterPlotter matrix, final int cx, final int cy, final int radius, final int angle, final String message) {
+    public static void arcPrint(final RasterPlotter matrix, final int cx, final int cy, final int radius, final double angle, final String message) {
         final int x = cx + (int) ((radius + 1) * Math.cos(RasterPlotter.PI180 * angle));
         final int y = cy - (int) ((radius + 1) * Math.sin(RasterPlotter.PI180 * angle));
         int yp = y + 3;
